+++ packages/google_mobile_ads/CHANGELOG.md
@@ -2,6 +2,11 @@
 * The minimum supported Flutter version is now 3.7.0.
 * Removes `visibility_detector` as a dependency, and the workaround added in
   https://github.com/googleads/googleads-mobile-flutter/pull/610.
+* Adds null checks for Ad Ids for Android in
+  https://github.com/googleads/googleads-mobile-flutter/pull/967
+* Updated Android dependencies in https://github.com/googleads/googleads-mobile-flutter/pull/843
+* Updates GMA iOS dependency to 10.11.0
+* Updates GMA Android dependency to 22.5.0
 
 ## 3.1.0
 * Updates GMA iOS dependency to 10.9.0